+++ pkg/artifactory/resource/webhook/resource_artifactory_webhook_build.go
@@ -15,8 +15,8 @@ type BuildWebhookCriteria struct {
 	SelectedBuilds []string `json:"selectedBuilds"`
 }
 
-var buildWebhookSchema = func(webhookType string) map[string]*schema.Schema {
-	return util.MergeSchema(baseWebhookBaseSchema(webhookType), map[string]*schema.Schema{
+var buildWebhookSchema = func(webhookType string, version int) map[string]*schema.Schema {
+	return util.MergeSchema(getBaseSchemaByVersion(webhookType, version), map[string]*schema.Schema{
 		"criteria": {
 			Type:     schema.TypeSet,
 			Required: true,