+++ pkg/artifactory/resource/webhook/resource_artifactory_webhook_repo.go
@@ -16,8 +16,8 @@ type RepoWebhookCriteria struct {
 	RepoKeys  []string `json:"repoKeys"`
 }
 
-var repoWebhookSchema = func(webhookType string) map[string]*schema.Schema {
-	return util.MergeSchema(baseWebhookBaseSchema(webhookType), map[string]*schema.Schema{
+var repoWebhookSchema = func(webhookType string, version int) map[string]*schema.Schema {
+	return util.MergeSchema(getBaseSchemaByVersion(webhookType, version), map[string]*schema.Schema{
 		"criteria": {
 			Type:     schema.TypeSet,
 			Required: true,